requirements.txt
matplotlib
librosa==0.8.0
tqdm
pandas
numba==0.48.0
numpy==1.20.0
scipy==1.5.4
# PyYAML==5.3.1
tensorboardX
pyloudnorm
setuptools>=41.0.0
g2p_en
resemblyzer==0.1.1.dev0
webrtcvad
tensorboard==2.6.0
scikit-learn==0.24.1
scikit-image==0.16.2
textgrid
jiwer
pycwt
PyWavelets
praat-parselmouth==0.3.3
jieba
einops
chardet
pretty-midi==0.2.9
pytorch-lightning==0.7.1
h5py==3.1.0
pypinyin==0.39.0
g2pM==0.1.2.5

TextGrid
pretty_midi
tensorboard
pycwt
pypinyin
protobuf==3.20
# crepe
lightning==2.1.0

# -i https://pypi.tuna.tsinghua.edu.cn/simple --upgrade pip

